refactor: move components to shared module
