+++ backend/user-service/src/main/java/org/linkwave/userservice/security/SecurityConfiguration.java
@@ -1,6 +1,7 @@
 package org.linkwave.userservice.security;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.linkwave.shared.auth.BearerAuthenticationConverter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
